refactor: transform goole sign in to layered arch
only the first step, callback will be done in the next commit
