+++ src/app/api/auth/controller/index.ts
@@ -1,9 +1,11 @@
 import { signOut } from "../useCase/signOut";
 import { register } from "./register";
 import { signInEmail } from "./signInEmail";
+import { signInGoogle } from "./signInGoogle";
 
 export default {
   register,
   signInEmail,
+  signInGoogle,
   signOut,
 };
