+++ src/app/api/auth/useCase/index.ts
@@ -1,9 +1,11 @@
 import { register } from "./register";
 import { signInEmail } from "./signInEmail";
+import { generateGoogleSignIn } from "./generateGoogleSignIn";
 import { signOut } from "./signOut";
 
 export default {
   register,
   signInEmail,
+  generateGoogleSignIn,
   signOut,
 };
